Connor: Finished rough version of implementation details in Dissertation.docx. Updated KMeans to run with my data rather than just the example. Need to finish the last part of gathering an equal amount of movies from each cluster and returning this to the edge node.

+++ DataCentre/DataCentreWebServer/DataCentreWebServer/MachineLearning/Movie.cs
@@ -7,12 +7,12 @@
         public int ID;
         public string Title;
         public int Year;
-        public float PercentageHorror;
-        public float PercentageComedy;
-        public float PercentageAction;
-        public float PercentageAdventure;
-        public float PercentageFantasy;
-        public float PercentageRomance;
+        public double PercentageHorror;
+        public double PercentageComedy;
+        public double PercentageAction;
+        public double PercentageAdventure;
+        public double PercentageFantasy;
+        public double PercentageRomance;
         public bool ContainsViolence;
         public bool ContainsSexualScenes;
         public bool ContainsDrugUse;
